ditch shapeless for the sake of simplicity (for now)
