+++ src/main/scala/tree/KDTree.scala
@@ -1,37 +1,23 @@
 package tree
 
-import distance.Distance.{Euclidean, Manhattan}
-import generic.GenericUtils.{reducerPoly, sqDiffMap, absDiffMap}
-import shapeless.ops.hlist.{LeftFolder, Mapper, Zip}
-import shapeless.{::, Generic, HList, HNil}
-import shapeless.ops.tuple.ToList
+import distance.Distance
 
 import scala.concurrent.{ExecutionContext, Future}
 
-sealed trait KDTree[T <: Manhattan[T]]{
+sealed trait KDTree[T <: Distance[T]]{
 
   val upperBoundaries: List[Option[BigDecimal]]
   val lowerBoundaries: List[Option[BigDecimal]]
 
-  def nnSearch[H<: HList, L<: HList](m: Int, target: T, currentBest: List[(T, BigDecimal)] = Nil)(
-    implicit gen: Generic.Aux[T, H],
-    zipper: Zip.Aux[H::H::HNil, L],
-    diffMapper: Mapper.Aux[absDiffMap.type, L, H],
-    folder: LeftFolder.Aux[H, BigDecimal, reducerPoly.type, BigDecimal]
-  ): List[(T, BigDecimal)]
+  def nnSearch(m: Int, target: T, currentBest: List[(T, BigDecimal)] = Nil): List[(T, BigDecimal)]
 //  def traverse(point: T): Leaf[T]
 }
 
-case class Leaf[T <: Manhattan[T]](point: T,
-                                   upperBoundaries: List[Option[BigDecimal]],
-                                   lowerBoundaries: List[Option[BigDecimal]]) extends KDTree[T]{
+case class Leaf[T <: Distance[T]](point: T,
+                                           upperBoundaries: List[Option[BigDecimal]],
+                                           lowerBoundaries: List[Option[BigDecimal]]) extends KDTree[T]{
 
-  override def nnSearch[H<: HList, L<: HList](m: Int, target: T, currentBest: List[(T, BigDecimal)] = Nil)(
-    implicit gen: Generic.Aux[T, H],
-    zipper: Zip.Aux[H::H::HNil, L],
-    diffMapper: Mapper.Aux[absDiffMap.type, L, H],
-    folder: LeftFolder.Aux[H, BigDecimal, reducerPoly.type, BigDecimal]
-  ): List[(T, BigDecimal)] = {
+  override def nnSearch(m: Int, target: T, currentBest: List[(T, BigDecimal)] = Nil): List[(T, BigDecimal)] = {
     val distanceToTarget: BigDecimal = target.distance(point)
 //   Take the m points with the shortest distances and then sort the list in descending order, so that the relevant neighbor is
 //    in the head of the list.
@@ -42,29 +28,22 @@ case class Leaf[T <: Manhattan[T]](point: T,
 //  override def traverse(point: T): Leaf[T] = this
 }
 
-case class Node[T <: Manhattan[T]](axis: Int,
-                   splitValue: BigDecimal,
-                   upperBoundaries: List[Option[BigDecimal]],
-                   lowerBoundaries: List[Option[BigDecimal]],
-                   left: KDTree[T],
-                   right: KDTree[T])(
-    implicit toList: ToList.Aux[T, BigDecimal, List[BigDecimal]]
-) extends KDTree[T]{
+case class Node[T <: Distance[T]](axis: Int,
+                                           splitValue: BigDecimal,
+                                           upperBoundaries: List[Option[BigDecimal]],
+                                           lowerBoundaries: List[Option[BigDecimal]],
+                                           left: KDTree[T],
+                                           right: KDTree[T]) extends KDTree[T]{
 
   import KDTree._
 
-  def nnSearch[H<: HList, L<: HList](m: Int,
+  def nnSearch(m: Int,
                target: T,
-               currentBest: List[(T, BigDecimal)] = Nil)(
-    implicit gen: Generic.Aux[T, H],
-    zipper: Zip.Aux[H::H::HNil, L],
-    diffMapper: Mapper.Aux[absDiffMap.type, L, H],
-    folder: LeftFolder.Aux[H, BigDecimal, reducerPoly.type, BigDecimal]
-  ): List[(T, BigDecimal)] = {
+               currentBest: List[(T, BigDecimal)] = Nil): List[(T, BigDecimal)] = {
 
 //  Search left first, then right
     if({
-      val asList = target.toList
+      val asList = target.features
       asList(axis) <= splitValue
     } ){
 
@@ -106,13 +85,12 @@ case class Node[T <: Manhattan[T]](axis: Int,
 
 object KDTree {
 
-  def grow[T <: Manhattan[T]](data: List[T],
-                              i: Int = 0,
-           upperBoundaries: List[Option[BigDecimal]] = Nil,
-           lowerBoundaries: List[Option[BigDecimal]] = Nil)
-          (
-            implicit ec: ExecutionContext,
-           toList: ToList.Aux[T, BigDecimal, List[BigDecimal]]
+  def grow[T <: Distance[T]](data: List[T],
+                                      i: Int = 0,
+                                      upperBoundaries: List[Option[BigDecimal]] = Nil,
+                                      lowerBoundaries: List[Option[BigDecimal]] = Nil)
+                                     (
+            implicit ec: ExecutionContext
           ): Future[KDTree[T]] ={
 
     val nextAxis: Int = getNextAxis(data, i)
@@ -120,8 +98,8 @@ object KDTree {
     val (filledLowerBoundaries, filledUpperBoundaries) = {
       if(upperBoundaries == Nil && lowerBoundaries == Nil){
         (
-          data.head.toList.map{_=> None},
-          data.head.toList.map{_=> None}
+          data.head.features.map{_=> None},
+          data.head.features.map{_=> None}
         )
       }
       else (lowerBoundaries, upperBoundaries)
@@ -135,17 +113,17 @@ object KDTree {
       )
       case l: List[T] =>
         val split = median(l, nextAxis){p: T =>
-          val listRepr: List[BigDecimal] = p.toList
+          val listRepr: List[BigDecimal] = p.features
           listRepr(nextAxis)
         }
 
         val leftData: List[T] = data.filter{p:T =>
-          val listRepr: List[BigDecimal] = p.toList
+          val listRepr: List[BigDecimal] = p.features
           listRepr(nextAxis) <= split
         }
 
         val rightData: List[T] = data.filter{p:T =>
-          val listRepr: List[BigDecimal] = p.toList
+          val listRepr: List[BigDecimal] = p.features
           listRepr(nextAxis) > split
         }
 
@@ -171,9 +149,7 @@ object KDTree {
     }
   }
 
-  def median[T <: Manhattan[T]](data: List[T], axis: Int)(mapper: T => BigDecimal)(
-    implicit toList: ToList.Aux[T, BigDecimal, List[BigDecimal]]
-  ): BigDecimal = {
+  def median[T <: Distance[T]](data: List[T], axis: Int)(mapper: T => BigDecimal): BigDecimal = {
     val mapped = data.sortBy(sortByAxis(axis)).par.map(mapper)
     if(mapped.length % 2 == 0){
       mapped(mapped.size/2 - 1) + (mapped(mapped.size - mapped.size/2) - mapped(mapped.size/2 - 1))/2.0
@@ -186,19 +162,17 @@ object KDTree {
     boundary.updated(dimension, Some(splitValue))
   }
 
-  def sortByAxis[T <: Manhattan[T]](currentAxis: Int)(implicit toList: ToList.Aux[T, BigDecimal, List[BigDecimal]]): T => BigDecimal = { p: T =>
-    val asList = p.toList
+  def sortByAxis[T <: Distance[T]](currentAxis: Int): T => BigDecimal = { p: T =>
+    val asList = p.features
     asList(currentAxis)
   }
 
-  def ballWithinBounds[T <: Manhattan[T]](
+  def ballWithinBounds[T <: Distance[T]](
                        m: Int,
                        query: T,
                        upperBound: List[Option[BigDecimal]],
                        lowerBound: List[Option[BigDecimal]],
-                       mNN: List[(T, BigDecimal)])(
-    implicit toList: ToList.Aux[T, BigDecimal, List[BigDecimal]]
-  ): Boolean = {
+                       mNN: List[(T, BigDecimal)]): Boolean = {
 
     mNN match{
       case _: List[(T, BigDecimal)] if mNN.size < m => false
@@ -209,7 +183,7 @@ object KDTree {
         val zipped = lowerBound.zip(upperBound).zipWithIndex
 
         val returns = zipped.foldLeft(false){case (res, ((l, u), dim))=>
-          val asList: List[BigDecimal] = query.toList
+          val asList: List[BigDecimal] = query.features
           val value: BigDecimal = asList(dim)
 
             res ||
@@ -222,19 +196,17 @@ object KDTree {
     }
   }
 
-  def boundsOverlapBall[T <: Manhattan[T]](
+  def boundsOverlapBall[T <: Distance[T]](
                         m: Int,
                         query: T,
                         upperBound: List[Option[BigDecimal]],
                         lowerBound: List[Option[BigDecimal]],
-                        mNN: List[(T, BigDecimal)])(
-    implicit toList: ToList.Aux[T, BigDecimal, List[BigDecimal]]
-  ): Boolean = {
+                        mNN: List[(T, BigDecimal)]): Boolean = {
 
     if(mNN.size < m) true
     else {
 
-        query.toList.zipWithIndex.foldLeft(BigDecimal(0)){case (cumSum, (v, i))=>
+        query.features.zipWithIndex.foldLeft(BigDecimal(0)){case (cumSum, (v, i))=>
           (lowerBound(i), upperBound(i)) match {
             case (Some(l), _) if v <= l =>
               val coordDist = cumSum + v.coordinateDistance(l)
@@ -247,50 +219,15 @@ object KDTree {
               else coordDist
             case _ => return true
           }
-//        if(lowerBound(i).exists(_ > v))
-//          lowerBound(i).map(v.coordinateDistance(_) + cumSum).getOrElse(cumSum)
-//        else if(upperBound(i).exists( _ < v)) upperBound(i).map(v.coordinateDistance(_) + cumSum).getOrElse(cumSum)
-//        else 0
       } > mNN.head._2
 
-//
-//
-//
-//      val sum: Option[BigDecimal] =  {
-//        if(lowerBound.head.exists(_ > query.x)) lowerBound.head.map(query.x.coordinateDistance)
-//        else if(upperBound.head.exists( _ < query.x)) upperBound.head.map(query.x.coordinateDistance)
-//        else None
-//      }.flatMap{cumSum =>
-//        if(lowerBound.tail.head.exists(_ > query.y)) lowerBound.tail.head.map(query.y.coordinateDistance).map(_ + cumSum)
-//        else if(upperBound.tail.head.exists(_ < query.y)) upperBound.tail.head.map(query.y.coordinateDistance).map(_ + cumSum)
-//        else None
-//      }
-//
-//      sum.forall{_ > mNN.head._2}
     }
   }
 
-  def getNextAxis[T <: Manhattan[T]](data: List[T], current: Int)(
-    implicit toList: ToList.Aux[T, BigDecimal, List[BigDecimal]]
-  ): Int = {
+  def getNextAxis[T <: Distance[T]](data: List[T], current: Int): Int = {
 
-    val maxDims = data.head.toList.size - 1
+    val maxDims = data.head.features.size - 1
     if(current == maxDims ) 0 else current + 1
-
-//    val (minVals: List[BigDecimal], maxVals: List[BigDecimal]) = data
-//      .tail
-//      .foldLeft[(List[BigDecimal], List[BigDecimal])]((data.head.toList, data.head.toList)){case (cumulative, next)=>
-//      (
-//        cumulative._1.zip(next.toList).map{case (x0, x1)=> x0.min(x1)},
-//        cumulative._2.zip(next.toList).map{case (x0, x1)=> x0.max(x1)}
-//      )
-//    }
-//
-//    minVals
-//      .zip(maxVals)
-//      .map{case (x, y)=> (x-y).abs}
-//      .zipWithIndex
-//      .maxBy{case (value, _)=> value}._2
   }
 
   implicit class Coordinate(t1: BigDecimal){
